% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Altitude.R
\name{Altitude}
\alias{Altitude}
\title{Calculate the altitude from pressure level in Pa.
 TODO: consider lon values}
\usage{
Altitude(p)
}
\arguments{
\item{p}{pressure in Pa/mbar}
}
\value{
vector with altitudes in m
}
\description{
Calculate the altitude from pressure level in Pa.
 TODO: consider lon values
}

