% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CDO.R
\name{CDO}
\alias{CDO}
\title{System call of the cdo binary.}
\usage{
CDO(..., sep = "", debug = FALSE)
}
\arguments{
\item{...}{one or more arguments of type string, which can be concatenated
using the paste function (e.g. '-yseasmean', '-fldmean').
Additionally, parameter for the system function can be passed, e.g.,
'intern = TRUE' to capture the output of the cdo command.}

\item{sep}{character used to seperate the terms. Used in the paste function.}

\item{debug}{print cdo command before execution and call cdo without '-s' argument.}
}
\description{
System call of the cdo binary.
}
\examples{
infile <- "~/data/u_50hPa_era40_DJF.grb"
     tmpfile <- tempfile()
     CDO('-fldmean -selyear,1980/1990',infile,tmpfile,sep=' ')
}

