% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetVarNCDF.R
\name{ConvertTime}
\alias{ConvertTime}
\title{Tries to convert date information, e.g, from netcdf file, to Date or Datetime object.}
\usage{
ConvertTime(time, unit)
}
\arguments{
\item{time}{vector containing the dates}

\item{unit}{string description the date format}
}
\value{
vector of Date or Datetime objects in case of success
}
\description{
Tries to convert date information, e.g, from netcdf file, to Date or Datetime object.
}
\examples{
ConvertTime(seq(0,1000), 'days since 1900-01-01 00:00:00')
}

