% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetrendSimple.R
\name{DetrendLinear}
\alias{DetrendLinear}
\title{Simple linear detrending of a vector. The detrending approach tries to 
remove a systematic drift in the data, which is found, for instances, sometimes
in climate model simulations.}
\usage{
DetrendLinear(d)
}
\arguments{
\item{d}{vector to detrend}
}
\value{
detrended vector
}
\description{
Simple linear detrending of a vector. The detrending approach tries to 
remove a systematic drift in the data, which is found, for instances, sometimes
in climate model simulations.
}
