% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetVarNCDF.R
\name{GetDimNCDF}
\alias{GetDimNCDF}
\title{Returns dimension of a netcdf variable.}
\usage{
GetDimNCDF(ncfile, varname)
}
\arguments{
\item{ncfile}{filename of the netcdf file.}

\item{varname}{name of the variable to extract.}
}
\value{
a vector of dimensions names.
}
\description{
Returns dimension of a netcdf variable.
}

