% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/GetVarNCDF.R
\name{GetVarNCDFL}
\alias{GetVarNCDFL}
\title{GetVarNCDFTL extract variable with one dimension of lat|lon|lev.}
\usage{
GetVarNCDFL(ncfile, varname)
}
\arguments{
\item{ncfile}{filename of the netcdf file}

\item{varname}{name of the variable to extract}
}
\value{
A data frame containing two columns.
}
\description{
GetVarNCDFTL extract variable with one dimension of lat|lon|lev.
}

