% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVarNCDF.R
\name{GetVarNCDFLL}
\alias{GetVarNCDFLL}
\title{GetVarNCDFLL extract variable with two dimensions of lat|lon|lev.}
\usage{
GetVarNCDFLL(ncfile, varname)
}
\arguments{
\item{ncfile}{filename of the netcdf file}

\item{varname}{name of the variable to extract}
}
\value{
A data frame containing three columns.
}
\description{
GetVarNCDFLL extract variable with two dimensions of lat|lon|lev.
}
