% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetVarNCDF.R
\name{GetVarNCDFLLL}
\alias{GetVarNCDFLLL}
\title{GetVarNCDFLLL extract variable with lat, lon, and lev dimension.}
\usage{
GetVarNCDFLLL(ncfile, varname)
}
\arguments{
\item{ncfile}{filename of the netcdf file}

\item{varname}{name of the variable to extract}
}
\value{
A data frame containing four columns.
}
\description{
GetVarNCDFLLL extract variable with lat, lon, and lev dimension.
}

