% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVarNCDF.R
\name{GetVarNCDFT}
\alias{GetVarNCDFT}
\title{GetVarNCDFT is a shortcut for GetVarNCDF(ncfile, varname, c('time')}
\usage{
GetVarNCDFT(ncfile, varname, convert_date = TRUE)
}
\arguments{
\item{ncfile}{filename of the netcdf file}

\item{varname}{name of the variable to extract}

\item{convert_date}{Convert time information into Date/Datetime object. Defaults to TRUE.}
}
\value{
A data frame containing two columns (date, value)
}
\description{
GetVarNCDFT is a shortcut for GetVarNCDF(ncfile, varname, c('time')
}
\examples{
     
     ncfile <- 'example_t.nc'
     d <- GetVarNCDF(ncfile, 'slp')


}
