% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/GetVarNCDF.R
\name{GetVarNCDFTL}
\alias{GetVarNCDFTL}
\title{GetVarNCDFTL extract variable with time dimension and one of lat|lon|lev.}
\usage{
GetVarNCDFTL(ncfile, varname, convert_date = TRUE)
}
\arguments{
\item{ncfile}{filename of the netcdf file}

\item{varname}{name of the variable to extract}
}
\value{
A data frame containing three columns.
}
\description{
GetVarNCDFTL extract variable with time dimension and one of lat|lon|lev.
}

