% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVarNCDF.R
\name{GetVarNCDFTLL}
\alias{GetVarNCDFTLL}
\title{GetVarNCDFTLL extract variable with time dimensions and two of lat|lon|lev.}
\usage{
GetVarNCDFTLL(ncfile, varname, convert_date = TRUE)
}
\arguments{
\item{ncfile}{filename of the netcdf file}

\item{varname}{name of the variable to extract}

\item{convert_date}{Convert time information into Date/Datetime object. Defaults to TRUE.}
}
\value{
A data frame containing four columns.
}
\description{
GetVarNCDFTLL extract variable with time dimensions and two of lat|lon|lev.
}
\examples{
     
     ncfile <- 'example_tll.nc'
     d <- GetVarNCDF(ncfile, 'slp')


}
