% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/GetVarNCDF.R
\name{GetVarNCDFTLLL}
\alias{GetVarNCDFTLLL}
\title{GetVarNCDFTLLL extract variable with time, lat, lon, and lev dimension.}
\usage{
GetVarNCDFTLLL(ncfile, varname, convert_date = TRUE)
}
\arguments{
\item{ncfile}{filename of the netcdf file}

\item{varname}{name of the variable to extract}
}
\value{
A data frame containing five columns.
}
\description{
GetVarNCDFTLLL extract variable with time, lat, lon, and lev dimension.
}

