% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/GlobalMean.R
\name{GlobalMean}
\alias{GlobalMean}
\title{Calculate weighted global mean from data.table. Weighting is done by latitudes.
 TODO: consider lon values}
\usage{
GlobalMean(d)
}
\arguments{
\item{d}{data table containing the column lat (latitude) and value (the variable)}
}
\value{
vector with global mean values.
}
\description{
Calculate weighted global mean from data.table. Weighting is done by latitudes.
 TODO: consider lon values
}

