% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LaggedCorrelation.R
\name{LaggedCorrelation}
\alias{LaggedCorrelation}
\title{Calculate pearson correlation coefficient between two data sets for different lags.}
\usage{
LaggedCorrelation(x, y, lags, independent.obs = length(x), parallel = T)
}
\arguments{
\item{x}{vector containing the independent variable}

\item{y}{vector containing the depended variable}

\item{lags}{vector of lag}

\item{independent.obs}{number of independent observations. Defaults to length(x), when x or y is smoothed, the number of independent observations may be lower.}

\item{parallel}{calculate in parallel, using the foreach package (default True)}
}
\value{
data frame containing the lag applied, the correlation coefficient (r), the p-value (p) and the number of independent observations used to calculate the p-value.
}
\description{
Calculate pearson correlation coefficient between two data sets for different lags.
}

