% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MyCorTest.R
\name{MyCorTest}
\alias{MyCorTest}
\title{Calculate p-value for a given correlation coefficient an estimate for the number of independent observations.}
\usage{
MyCorTest(r, n, conf.level = NULL, alternative = "two.sided")
}
\arguments{
\item{r}{correlation coefficient}

\item{n}{number of independent observations}

\item{conf.level}{(optional) test estimated p value against a given confidence level. Returns True or False.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of ‘"two.sided"’ (default), ‘"greater"’ or ‘"less"’. So far, only two.sided is implemented.}
}
\value{
p value or True/False, when conf.level is specified.
}
\description{
Calculate p-value for a given correlation coefficient an estimate for the number of independent observations.
}

