% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/MyTtest.R
\name{MyTtest}
\alias{MyTtest}
\title{Calculate p value Of Student's t-test for significant differences based on mean value,
standard deviation, and number of independend observations.}
\usage{
MyTtest(m1, s1, n1, m2, s2, n2)
}
\arguments{
\item{m1}{mean value of first data set}

\item{s1}{standard deviation of first data set}

\item{n1}{number of (independed) observations of first data set.}

\item{m2}{mean value of second data set}

\item{s2}{standard deviation of second data set}

\item{n2}{number of (independed) observations of second data set.}
}
\value{
p value.
}
\description{
Calculate p value Of Student's t-test for significant differences based on mean value,
standard deviation, and number of independend observations.
}
\examples{
d1 = rnorm(100)
 d2 = rnorm(100)

 MyTtest(mean(d1), sd(d1), length(d1), mean(d2), sd(d2), length(d2))
}

