% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/SeasonalMean.R
\name{SeasonalMean}
\alias{SeasonalMean}
\title{Calculate seasonal means from data.table}
\usage{
SeasonalMean(data, season, na.rm = FALSE)
}
\arguments{
\item{data}{data table containing the columns date (Date object) and value (the variable)}

\item{season}{string descripting the season to average, e.g., 'DJF'}

\item{na.rm}{a logical value indicating whether 'NA' values should be
stripped before the computation proceeds.}
}
\value{
data table with seasonal averaged values.
}
\description{
Calculate seasonal means from data.table
}
\examples{
mydata <- data.frame(date = seq.Date(as.Date('1990-01-01'),
                                  as.Date('2000-12-31'),
                                  by = '1 months'),
                                  value = rnorm(11*12))
SeasonalMean(mydata, 'DJF')
SeasonalMean(mydata, 'DJFMAM')
}

