% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SystemCmd.R
\name{SystemCmd}
\alias{SystemCmd}
\title{Wrapper function to system()}
\usage{
SystemCmd(..., sep = "", debug = FALSE, intern = FALSE,
  ignore.stdout = FALSE, ignore.stderr = FALSE, wait = TRUE,
  input = NULL)
}
\arguments{
\item{...}{one or more arguments of type string, which can be concatenated
using the paste function (e.g. 'ls','-lrt','/bin').}

\item{sep}{character used to seperate the terms. Used in the paste function.}

\item{debug}{print command before execution.}

\item{intern}{a logical (not 'NA') which indicates whether to capture the
output of the command as an R character vector (parameter to system
function).}

\item{ignore.stdout,ignore.stderr}{a logical (not 'NA') indicating whether
messages written to 'stdout' or 'stderr' should be ignored (parameter
to system function).}

\item{wait}{a logical (not 'NA') indicating whether the R interpreter
should wait for the command to finish, or run it
asynchronously.  This will be ignored (and the interpreter
will always wait) if 'intern = TRUE' (parameter to system function).}

\item{input}{if a character vector is supplied, this is copied one string
per line to a temporary file, and the standard input of
'command' is redirected to the file (parameter to system function).}
}
\value{
Return value of the executed command.
}
\description{
Wrapper function to system()
}
\examples{
SystemCmd('ls -lrt /')
    SystemCmd('ls -lrt /', debug = TRUE)
    SystemCmd('ls', '-lrt', '/')
}

