% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/YearMean.R
\name{YearMean}
\alias{YearMean}
\title{Calculate annual means from data.table}
\usage{
YearMean(data, varname = "value", na.rm = TRUE)
}
\arguments{
\item{data}{data table containing the columns date (Date object) and value (the variable)}

\item{varname}{name of the variable, defaults to 'value''}

\item{na.rm}{a logical value indicating whether 'NA' values should be
stripped before the computation proceeds. Defaults to TRUE}
}
\value{
data table with seasonal averaged values.
}
\description{
Calculate annual means from data.table
}
\examples{
mydata <- data.frame(date = seq.Date(as.Date('1990-01-01'),
                                  as.Date('2000-12-31'),
                                  by = '1 months'),
                                  value = rnorm(11*12))
YearMean(mydata,)
}

