% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVarNCDF.R
\name{GetVarNCDF}
\alias{GetVarNCDF}
\title{Extracts data from netcdf file and returns a data.frame.}
\usage{
GetVarNCDF(ncfile, varname, finddim, convert_date = TRUE)
}
\arguments{
\item{ncfile}{filename of the netcdf file.}

\item{varname}{name of the variable to extract.}

\item{finddim}{vector of dimension variables, e.g, c('time','lat','lev'). 
For dimensionless variables, e.g, global mean values, and empty vector or NULL
should be used.}

\item{convert_date}{Convert time information into Date/Datetime object. Defaults to TRUE.
This parameter is only used, when a date/time variable is found.}
}
\value{
A data frame containing at least a single column (value) for the variable value(s) 
        and possible additional columns for each dimension variable.
}
\description{
Extracts data from netcdf file and returns a data.frame.
}
\examples{
     
     ncfile <- 'example_tll.nc'
     d <- GetVarNCDF(ncfile, 'slp', c('lon','lat','time'))

}
