% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVarNCDF.R
\name{GetVarNCDFTLLL}
\alias{GetVarNCDFTLLL}
\title{GetVarNCDFTLLL extract variable with time, lat, lon, and lev dimension.}
\usage{
GetVarNCDFTLLL(ncfile, varname, convert_date = TRUE)
}
\arguments{
\item{ncfile}{filename of the netcdf file}

\item{varname}{name of the variable to extract}

\item{convert_date}{Convert time information into Date/Datetime object. Defaults to TRUE.
This parameter is only used, when a date/time variable is found.}
}
\value{
A data frame containing five columns.
}
\description{
GetVarNCDFTLLL extract variable with time, lat, lon, and lev dimension.
}
